+++ source/extensions/isaacsim.core.experimental.utils/python/mock/xform.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2021-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,15 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import omni.ext
-
-from .rendering_manager import RenderingEvent, RenderingManager
-from .viewport_manager import ViewportManager
-
-
-class Extension(omni.ext.IExt):
-    def on_startup(self, ext_id):
-        pass
-
-    def on_shutdown(self):
-        RenderingManager.deregister_all_callbacks()
+from .impl.xform import *