+++ source/extensions/isaacsim.core.rendering_manager/docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.0] - 2025-11-21
+### Added
+- Add method to set a camera view
+
+### Changed
+- Split the manager implementation into two: `RenderingManager` and `ViewportManager`
+
 ## [0.1.0] - 2025-11-11
 ### Added
 - Initial release